F_T32245 move SurveyBundle Logic to core
- adjust use statements
