+++ demosplan/DemosPlanCoreBundle/Controller/Survey/SurveyController.php
@@ -15,11 +15,11 @@ use demosplan\DemosPlanCoreBundle\Controller\Base\BaseController;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
 use demosplan\DemosPlanCoreBundle\Exception\MessageBagException;
 use demosplan\DemosPlanCoreBundle\Exception\SurveyInputDataException;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyCreateHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyNewHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyUpdateHandler;
 use demosplan\DemosPlanProcedureBundle\Logic\ProcedureHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyCreateHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyNewHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyUpdateHandler;
 use demosplan\DemosPlanSurveyBundle\Validator\SurveyValidator;
 use Exception;
 use Symfony\Component\HttpFoundation\Request;
