+++ demosplan/DemosPlanCoreBundle/Controller/Survey/SurveyVoteAPIController.php
@@ -20,13 +20,13 @@ use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\SurveyVote;
 use demosplan\DemosPlanCoreBundle\Exception\BadRequestException;
 use demosplan\DemosPlanCoreBundle\Exception\MessageBagException;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyService;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteCreateHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteService;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\SurveyVoteResourceType;
 use demosplan\DemosPlanProcedureBundle\Logic\CurrentProcedureService;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyService;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyVoteCreateHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyVoteHandler;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyVoteService;
 use demosplan\DemosPlanSurveyBundle\Validator\SurveyVoteValidator;
 use Exception;
 use Symfony\Component\HttpFoundation\Request;
