+++ tests/backend/core/Survey/Unit/SurveyCreateHandlerTest.php
@@ -12,7 +12,7 @@ namespace Tests\Core\Survey\Unit;
 
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
 use demosplan\DemosPlanCoreBundle\Exception\SurveyInputDataException;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyCreateHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyCreateHandler;
 use Exception;
 
 class SurveyCreateHandlerTest extends SurveyTestUtils
