+++ tests/backend/core/Survey/Unit/SurveyHandlerTest.php
@@ -12,7 +12,7 @@ namespace Tests\Core\Survey\Unit;
 
 use demosplan\DemosPlanCoreBundle\DataFixtures\ORM\TestData\LoadSurveyData;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyHandler;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyHandler;
 use Exception;
 
 class SurveyHandlerTest extends SurveyTestUtils
