+++ tests/backend/core/Survey/Unit/SurveyTestUtils.php
@@ -16,8 +16,8 @@ use demosplan\DemosPlanCoreBundle\DataFixtures\ORM\TestData\LoadSurveyData;
 use demosplan\DemosPlanCoreBundle\Entity\Procedure\Procedure;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\Survey;
 use demosplan\DemosPlanCoreBundle\Entity\Survey\SurveyVote;
+use demosplan\DemosPlanCoreBundle\Logic\Survey\SurveyVoteHandler;
 use demosplan\DemosPlanCoreBundle\Permissions\PermissionsInterface;
-use demosplan\DemosPlanSurveyBundle\Logic\SurveyVoteHandler;
 use Exception;
 use Symfony\Bridge\Monolog\Logger;
 use Tests\Base\UnitTestCase;
